+++ html-conduit/tagstream-conduit/Text/HTML/TagStream/Text.hs
@@ -1,8 +1,9 @@
 {-# LANGUAGE OverloadedStrings, TupleSections, ViewPatterns #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE RecordWildCards #-}
 module Text.HTML.TagStream.Text
-  ( Token' (..)
+  ( Token (..)
   , tokenStream
   ) where
 
@@ -17,11 +18,10 @@ import           Prelude hiding (mapM)
 import qualified Data.Attoparsec.ByteString.Char8 as S
 import           Data.Attoparsec.Text
 import           Data.ByteString (ByteString)
-import qualified Data.CaseInsensitive as CI
 import           Data.Conduit
 import           Data.Functor.Identity (runIdentity)
-import           Data.Maybe (fromMaybe)
-import           Data.Monoid (mconcat)
+import           Data.Maybe (fromMaybe, isJust)
+import           Data.Monoid (mconcat, (<>))
 import           Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.Lazy as L
@@ -35,13 +35,24 @@ import           Data.Conduit.Internal (pipeL)
 import qualified Data.Conduit.List as C
 import qualified Data.Conduit.Attoparsec as C
 import qualified Data.Conduit.Text as C
+import           Data.String (IsString)
+import           Data.Map (Map)
+import qualified Data.Map.Strict as Map
+import           Control.Arrow (first, second)
 
-import           Text.HTML.TagStream.Entities
-import           Text.HTML.TagStream.Types
-import           Text.HTML.TagStream.Utils (splitAccum)
+data Token
+  = TagOpen Text (Map Text Text) Bool
+  | TagClose Text
+  | Text Text
+  | Comment Text
+  | Special Text Text
+  | Incomplete Text
+  deriving (Eq, Show)
 
-type Token = Token' Text
-type Attr = Attr' Text
+data TagType
+  = TagTypeClose
+  | TagTypeSpecial
+  | TagTypeNormal
 
 {--
  - match quoted string, can fail.
@@ -80,7 +91,7 @@ tagEnd = char '>' *> pure False
 {--
  - attribute pair or tag end, can fail if tag end met.
  -}
-attr :: Parser Attr
+attr :: Parser (Text, Text)
 attr = (,) <$> attrName <* skipSpace
            <*> ( boolP (char '=') >>=
                  cond (skipSpace *> attrValue)
@@ -90,13 +101,13 @@ attr = (,) <$> attrName <* skipSpace
 {--
  - all attributes before tag end. can't fail.
  -}
-attrs :: Parser ([Attr], Bool)
-attrs = loop []
+attrs :: Parser (Map Text Text, Bool)
+attrs = loop Map.empty
   where
     loop acc = skipSpace *> (Left <$> tagEnd <|> Right <$> attr) >>=
                either
-                 (return . (reverse acc,))
-                 (loop . (:acc))
+                 (return . (acc,))
+                 (\(key, value) -> loop $ Map.insert key value acc)
 
 {--
  - comment tag without prefix.
@@ -225,16 +236,16 @@ maybeP p = Just <$> p <|> return Nothing
 -- }}}
 
 -- {{{ encode tokens
-cc :: [Text] -> B.Builder
+cc :: [Text] -> B.Builder -- FIXME use Builder correctly
 cc = mconcat . map B.fromText
 
 showToken :: (Text -> Text) -> Token -> B.Builder
 showToken hl (TagOpen name as close) =
     cc $ [hl "<", name]
-      ++ map showAttr as
+      ++ map showAttr (Map.toList as)
       ++ [hl (if close then "/>" else ">")]
   where
-    showAttr :: Attr -> Text
+    showAttr :: (Text, Text) -> Text
     showAttr (key, value) = T.concat $ [" ", key, hl "=\""] ++ map escape (T.unpack value) ++ [hl "\""]
     escape '"' = "\\\""
     escape '\\' = "\\\\"
@@ -270,3 +281,133 @@ tokenStream =
     close s r = do
         unless (T.null s) $ yield $ Text s
         return r
+
+splitAccum :: [Token] -> (Text, [Token])
+splitAccum [] = (mempty, [])
+splitAccum (reverse -> (Incomplete s : xs)) = (s, reverse xs)
+splitAccum tokens = (mempty, tokens)
+
+-- Entities
+
+-- | A conduit to decode entities from a stream of tokens into a new stream of tokens.
+decodeEntities :: (Monad m
+                  ,Monoid builder)
+               => Dec builder Text
+               -> Conduit Token m Token
+decodeEntities dec =
+    start
+  where
+    start = await >>= maybe (return ()) (\token -> start' token >> start)
+    start' (Text t) = (yield t >> yieldWhileText) =$= decodeEntities' dec =$= CL.mapMaybe go
+    start' (TagOpen name attrs bool) = yield (TagOpen name (Map.map (decodeString dec) attrs) bool)
+    start' token = yield token
+
+    go t
+        | t == ""   = Nothing
+        | otherwise = Just (Text t)
+
+-- | Decode entities in a complete string.
+decodeString
+  :: (Eq a, IsString a, Monoid builder, Monoid a)
+  => Dec builder a -> a -> a
+decodeString dec input =
+  case makeEntityDecoder dec input of
+    (value', remainder)
+      | value' /= mempty -> value' <> decodeString dec remainder
+      | otherwise -> input
+
+decodeEntities' :: (Monad m
+                   ,Monoid string
+                   ,IsString string
+                   ,Monoid builder
+                   ,Eq string)
+                => Dec builder string
+                -> Conduit string m string
+decodeEntities' dec =
+    loop id
+  where
+    loop accum = do
+        mchunk <- await
+        let chunk = accum $ fromMaybe mempty mchunk
+            (newStr, remainder) = makeEntityDecoder dec chunk
+        yield newStr
+        if isJust mchunk
+            then loop (mappend remainder)
+            else yield remainder
+
+-- | Yield contiguous text tokens as strings.
+yieldWhileText :: Monad m => Conduit Token m Text
+yieldWhileText =
+    loop
+  where
+    loop = await >>= maybe (return ()) go
+    go (Text t) = yield t >> loop
+    go token = leftover token
+
+-- | A decoder.
+data Dec builder string = Dec
+  { decToS     :: builder -> string
+  , decBreak   :: (Char -> Bool) -> string -> (string,string)
+  , decBuilder :: string -> builder
+  , decDrop    :: Int -> string -> string
+  , decEntity  :: string -> Maybe string
+  , decUncons  :: string -> Maybe (Char,string)
+  }
+
+-- | Decode the entities in a string type with a decoder.
+makeEntityDecoder :: (IsString string,Monoid builder,Eq string,Monoid string)
+                  => Dec builder string -> string -> (string, string)
+makeEntityDecoder Dec{..} = first decToS . go
+  where
+    go s =
+      case decBreak (=='&') s of
+        (_,"") -> (decBuilder s, "")
+        (before,restPlusAmp@(decDrop 1 -> rest)) ->
+          case decBreak (not . (\c -> isNameChar c || c == '#')) rest of
+            (_,"") -> (decBuilder before, restPlusAmp)
+            (entity,after) -> (before1 <> before2, after')
+              where
+                before1 = decBuilder before
+                (before2, after') =
+                  case mdecoded of
+                    Nothing -> first ((decBuilder "&" <> decBuilder entity) <>) (go after)
+                    Just (decBuilder -> decoded) ->
+                      case decUncons after of
+                        Just (';',validAfter) -> first (decoded <>) (go validAfter)
+                        Just (_invalid,_rest) -> first (decoded <>) (go after)
+                        Nothing -> (mempty, s)
+                mdecoded =
+                  if entity == mempty
+                     then Nothing
+                     else decEntity entity
+
+-- | Is the character a valid Name starter?
+isNameStart :: Char -> Bool
+isNameStart c =
+  c == ':' ||
+  c == '_' ||
+  isAsciiUpper c ||
+  isAsciiLower c ||
+  (c >= '\xC0' && c <= '\xD6') ||
+  (c >= '\xD8' && c <= '\xF6') ||
+  (c >= '\xF8' && c <= '\x2FF') ||
+  (c >= '\x370' && c <= '\x37D') ||
+  (c >= '\x37F' && c <= '\x1FFF') ||
+  (c >= '\x200C' && c <= '\x200D') ||
+  (c >= '\x2070' && c <= '\x218F') ||
+  (c >= '\x2C00' && c <= '\x2FEF') ||
+  (c >= '\x3001' && c <= '\xD7FF') ||
+  (c >= '\xF900' && c <= '\xFDCF') ||
+  (c >= '\xFDF0' && c <= '\xFFFD') ||
+  (c >= '\x10000' && c <= '\xEFFFF')
+
+-- | Is the character valid in a Name?
+isNameChar :: Char -> Bool
+isNameChar c =
+  c == '-' ||
+  c == '.' ||
+  c == '\xB7' ||
+  isDigit c ||
+  isNameStart c ||
+  (c >= '\x0300' && c <= '\x036F') ||
+  (c >= '\x203F' && c <= '\x2040')
